feature/cast-offer update duration type

+++ database/migrations/2019_05_20_122400_create_cast_offers_table.php
@@ -22,7 +22,7 @@ class CreateCastOffersTable extends Migration
             $table->string('address');
             $table->date('date');
             $table->time('start_time');
-            $table->float('duration');
+            $table->integer('duration');
             $table->integer('cost');
             $table->integer('temp_point');
             $table->tinyInteger('status')->default(1);
